refactor: 패키지, 클래스 이름 변경 User -> Member
